EY-2442 Flytt filtreringsting ut i egen fil
